zuul-core: fix SSL Bug
In newer versions of Netty, failure to write the ServerHello message result in a SslCompletionEvent and the channel closed. This can happen if the client sends a ClientHello, and then quickly closes the connection.

This change lowers the severity of the message being logged, and adds a test proving the connection cleans up. This test fails on older versions of netty.