zuul-core Add Cause of SSL Handshake Error from Origin
When SSL handler reads the message that the handshake failed (for example, because peer certificates are not verified), it populates the through handlers and eventually the failure message goes away. When ProxyEndpoint later tries to talk to origin it usually gets the error of SSLEngine Closed already, which does not display the actual reason why it was closed. This change catches the detailed failure cause and adds it to proxy attempt so client can know what went wrong:

"error":"General OpenSslEngine problem","cause":"java.security.cert.CertificateException: Certificate application name (applicationname) does not match expected name: expectedname","exceptionType":"SSLHandshakeException"