zuul-core: wire socket address from ClientChannelManager down to ClientConnectFactory
This is needed to ensure IPv6 support, and additionally to pass in LocalAddresses for testing in the future.

This avoids passing a string host around, and forces name resolution to happen earlier in the connection. The address to connect to is no longer owned by the Server object, and instead part of the connection pool. The Server object now functions as a way to logically reference a backend.

Additionally, several arguments are no longer used and have been removed, like the path, method, and attempt number in multiple places.

Lastly, blocking DNS reads are still possible, but this change is complicated as it is. I will make this async in a future PR.