zuul-core: set the selected address as a ip address string
Some of our smoke tests depend on the exact string form of the IP address, which was accidentally being set to `/127.0.0.1:1234`, rather than `127.0.0.1`.   I had commented on this originally since it was unclear what purpose it had, but now we know.


I'll try adding tests for this in a followup PR, but right now the code is not structured properly for tests.