zuul-core: refactor Filter loading into mutable and immutable parts
This step is needed to support the construction of static filter registries.  While
dynamic loading will still be supported, Zuul will be moving towards a fixed filter
set model.   This lowers load on live running servers, avoids a dependency on doing
disk reads.   Wide deploying code to live running servers is a risky endeavor, and
Zuul did not develop the necessary safeguards to support this.

After this change, FilterFileManager will be optional, as its sole purpose is to
populate the loader/registry.