zuul-processor: Make a Configuration based processor, and stop producing generated Java files. (#791)
While Gradle worked fine with the existing code, IntelliJ does not.  IntelliJ's
incremental compilation does not include the full classpath, making most of the
files in the Filter class not available.  The generated file thus changes frequently
and prevents running from the IDE.   JetBrains considers this a feature, and will
not change it.

Work around this by generating a list of classes, which is appended, rather than
created from scratch each compilation.  This allows incremental compilation in
IntelliJ to keep running at the cost of potentially having stale filters listed.