zuul-core: add a connection timer class (#878)
I picked to make a new Attrs class because the channel one doesn't really allow grouping. It also forces thread safety when we dont need it, and makes keys identical based on name. These have proved problematic in the past, so I made a new one.

The idea of the conn timer is to store it on the channel itself, along with the set of dimensions to break down stats by. The attrs map is dynamic, and allows updating it as new metrics come in. I plan on using these dimensions for other connection stats in the future.