zuul-core: add debugging and locking around CurrentPassport.
This is needed to help debug a seemingly impossible ConcurrentModificationException.   The passport should live entirely on a single thread, and not be re-entrant.  This code checks not only if the passport is locked, but also if it accidentally owns it already.   The lock will keep access safe, but it does not stop the sharing of the object. Once we know where the object is being accessed, we can disable this code.