Ensure SessionContext isn't modified when iterating over it
When routing debugging is enabled, Zuul will record changes in the
session context after each filter is run.  If the context already has a
"routingDebug" property on it, all will be fine.  If not, the comparison
code will (as a side effect) add that property.  However, it's added
while there is an active iterator over the context.  Once the code tries
to fetch the next element from the iterator, a
ConcurrentModificationException is thrown.