Extract client ip method (#1929)
There are cases where the client ip set in the http request object is
not correct. We should use the ATTR_CLIENT_IP value set in
ClientIpChannelHandler instead, which takes into account ppv2 and XFF
headers.
The only issue is that this value is scoped to edgezuul-netty, but the
http request is created in zuul oss in ClientRequestReceiver, which
means it is not accessible.
To deal with this, we can extract the code to get the client ip and then
override it in edgezuul-netty by creating an EdgeClientRequestReceiver
class that returns the client ip using
ClientIpChannelHandler.ATTR_CLIENT_IP instead. That PR will follow this
one.