[CGW-2515] Improve HttpHeadersTimeoutHandler event handling (#2017)
With this change we should free the timer future once the channel
becomes inactive. This change also removes the channel attribute when
the timeout is not reached so that we don't need to keep it around for
the lifetime of the channel.

This copies the event handling approach from
https://github.com/netty/netty/blob/fd85af2407742e9821ae77de598a887ba15b0082/handler/src/main/java/io/netty/handler/timeout/IdleStateHandler.java.

Unfortunately we can't quite inherit from Netty's IdleStateHandler or
ReadTimeoutHandler as those implement idle timeouts, not absolute
timeouts like this handler.